fix the signed unsigned comparison

+++ main.cpp
@@ -1,7 +1,7 @@
 #include "MapRecoloring.cpp"
 
 template<class T> void getVector(vector<T>& v) {
-    for (int i = 0; i < v.size(); ++i)
+    for (int i = 0; i < (int)v.size(); ++i)
         cin >> v[i];
 }
